[#1929] Reorganize advancement file structure to match other documents & applications
