--- module/advancement/_module.mjs
@@ -1,9 +1,0 @@
-export * as types from "./types/_module.mjs";
-
-export {default as Advancement} from "./advancement.mjs";
-export {default as AdvancementConfig} from "./advancement-config.mjs";
-export {default as AdvancementConfirmationDialog} from "./advancement-confirmation-dialog.mjs";
-export {default as AdvancementFlow} from "./advancement-flow.mjs";
-export {default as AdvancementManager} from "./advancement-manager.mjs";
-export {default as AdvancementSelection} from "./advancement-selection.mjs";
-
