--- module/advancement/types/_module.mjs
@@ -1,3 +1,0 @@
-export * from "./hit-points.mjs";
-export * from "./item-grant.mjs";
-export * from "./scale-value.mjs";
