+++ module/applications/_module.mjs
@@ -1,4 +1,5 @@
 export * as actor from "./actor/_module.mjs";
+export * as advancement from "./advancement/_module.mjs";
 export * as item from "./item/_module.mjs";
 export * as journal from "./journal/_module.mjs";
