+++ module/applications/advancement/advancement-confirmation-dialog.mjs
