+++ module/applications/advancement/item-grant-flow.mjs
@@ -1,0 +1,68 @@
+import AdvancementFlow from "./advancement-flow.mjs";
+
+/**
+ * Inline application that presents the player with a list of items to be added.
+ */
+export default class ItemGrantFlow extends AdvancementFlow {
+
+  /** @inheritdoc */
+  static get defaultOptions() {
+    return foundry.utils.mergeObject(super.defaultOptions, {
+      template: "systems/dnd5e/templates/advancement/item-grant-flow.hbs"
+    });
+  }
+
+  /* -------------------------------------------- */
+
+  /** @inheritdoc */
+  async getData() {
+    const config = this.advancement.configuration.items;
+    const added = this.retainedData?.items.map(i => foundry.utils.getProperty(i, "flags.dnd5e.sourceId"))
+      ?? this.advancement.value.added;
+    const checked = new Set(Object.values(added ?? {}));
+
+    const items = await Promise.all(config.map(fromUuid));
+    return foundry.utils.mergeObject(super.getData(), {
+      optional: this.advancement.configuration.optional,
+      items: items.reduce((arr, item) => {
+        if ( !item ) return arr;
+        item.checked = added ? checked.has(item.uuid) : true;
+        arr.push(item);
+        return arr;
+      }, [])
+    });
+  }
+
+  /* -------------------------------------------- */
+
+  /** @inheritdoc */
+  activateListeners(html) {
+    super.activateListeners(html);
+    html.find("a[data-uuid]").click(this._onClickFeature.bind(this));
+  }
+
+  /* -------------------------------------------- */
+
+  /**
+   * Handle clicking on a feature during item grant to preview the feature.
+   * @param {MouseEvent} event  The triggering event.
+   * @protected
+   */
+  async _onClickFeature(event) {
+    event.preventDefault();
+    const uuid = event.currentTarget.dataset.uuid;
+    const item = await fromUuid(uuid);
+    item?.sheet.render(true);
+  }
+
+  /* -------------------------------------------- */
+
+  /** @inheritdoc */
+  async _updateObject(event, formData) {
+    const retainedData = this.retainedData?.items.reduce((obj, i) => {
+      obj[foundry.utils.getProperty(i, "flags.dnd5e.sourceId")] = i;
+      return obj;
+    }, {});
+    await this.advancement.apply(this.level, formData, retainedData);
+  }
+}
