+++ module/applications/advancement/scale-value-config.mjs
@@ -1,110 +1,10 @@
-import Advancement from "../advancement.mjs";
-import AdvancementConfig from "../advancement-config.mjs";
-import AdvancementFlow from "../advancement-flow.mjs";
-
-import { ScaleValueConfigurationData, TYPES } from "../../data/advancement/scale-value.mjs";
-
-/**
- * Advancement that represents a value that scales with class level. **Can only be added to classes or subclasses.**
- */
-export class ScaleValueAdvancement extends Advancement {
-
-  /** @inheritdoc */
-  static get metadata() {
-    return foundry.utils.mergeObject(super.metadata, {
-      dataModels: {
-        configuration: ScaleValueConfigurationData
-      },
-      order: 60,
-      icon: "systems/dnd5e/icons/svg/scale-value.svg",
-      title: game.i18n.localize("DND5E.AdvancementScaleValueTitle"),
-      hint: game.i18n.localize("DND5E.AdvancementScaleValueHint"),
-      multiLevel: true,
-      validItemTypes: new Set(["class", "subclass"]),
-      apps: {
-        config: ScaleValueConfig,
-        flow: ScaleValueFlow
-      }
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * The available types of scaling value.
-   * @enum {ScaleValueType}
-   */
-  static TYPES = TYPES;
-
-  /* -------------------------------------------- */
-  /*  Instance Properties                         */
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  get levels() {
-    return Array.from(Object.keys(this.configuration.scale).map(l => Number(l)));
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Identifier for this scale value, either manual value or the slugified title.
-   * @type {string}
-   */
-  get identifier() {
-    return this.configuration.identifier || this.title.slugify();
-  }
-
-  /* -------------------------------------------- */
-  /*  Display Methods                             */
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  titleForLevel(level, { configMode=false }={}) {
-    const value = this.valueForLevel(level)?.display;
-    if ( !value ) return this.title;
-    return `${this.title}: <strong>${value}</strong>`;
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Scale value for the given level.
-   * @param {number} level      Level for which to get the scale value.
-   * @returns {ScaleValueType}  Scale value at the given level or null if none exists.
-   */
-  valueForLevel(level) {
-    const key = Object.keys(this.configuration.scale).reverse().find(l => Number(l) <= level);
-    const data = this.configuration.scale[key];
-    const TypeClass = this.constructor.TYPES[this.configuration.type];
-    if ( !data || !TypeClass ) return null;
-    return new TypeClass(data, { parent: this });
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Compare two scaling values and determine if they are equal.
-   * @param {*} a
-   * @param {*} b
-   * @returns {boolean}
-   */
-  testEquality(a, b) {
-    const keys = Object.keys(a ?? {});
-    if ( keys.length !== Object.keys(b ?? {}).length ) return false;
-    for ( const k of keys ) {
-      if ( a[k] !== b[k] ) return false;
-    }
-    return true;
-  }
-
-}
-
+import AdvancementConfig from "./advancement-config.mjs";
+import { TYPES } from "../../data/advancement/scale-value.mjs";
 
 /**
  * Configuration application for scale values.
  */
-export class ScaleValueConfig extends AdvancementConfig {
+export default class ScaleValueConfig extends AdvancementConfig {
 
   /** @inheritdoc */
   static get defaultOptions() {
@@ -120,14 +20,14 @@ export class ScaleValueConfig extends AdvancementConfig {
   /** @inheritdoc */
   getData() {
     const config = this.advancement.configuration;
-    const type = ScaleValueAdvancement.TYPES[config.type];
+    const type = TYPES[config.type];
     return foundry.utils.mergeObject(super.getData(), {
       classIdentifier: this.item.identifier,
       previewIdentifier: config.identifier || this.advancement.title?.slugify()
         || this.advancement.constructor.metadata.title.slugify(),
       type: type.metadata,
       types: Object.fromEntries(
-        Object.entries(ScaleValueAdvancement.TYPES).map(([key, d]) => [key, game.i18n.localize(d.metadata.label)])
+        Object.entries(TYPES).map(([key, d]) => [key, game.i18n.localize(d.metadata.label)])
       ),
       faces: Object.fromEntries(TYPES.dice.FACES.map(die => [die, `d${die}`])),
       levels: this._prepareLevelData(),
@@ -259,8 +159,8 @@ export class ScaleValueConfig extends AdvancementConfig {
         return obj;
       }, {}));
       updates.configuration.scale ??= {};
-      const OriginalType = ScaleValueAdvancement.TYPES[this.advancement.configuration.type];
-      const NewType = ScaleValueAdvancement.TYPES[updates.configuration.type];
+      const OriginalType = TYPES[this.advancement.configuration.type];
+      const NewType = TYPES[updates.configuration.type];
       for ( const [lvl, data] of Object.entries(updates.configuration.scale) ) {
         const original = new OriginalType(data, { parent: this.advancement });
         updates.configuration.scale[lvl] = NewType.convertFrom(original)?.toObject();
@@ -268,29 +168,4 @@ export class ScaleValueConfig extends AdvancementConfig {
     }
     return super._updateObject(event, foundry.utils.flattenObject(updates));
   }
-}
-
-
-/**
- * Inline application that displays any changes to a scale value.
- */
-export class ScaleValueFlow extends AdvancementFlow {
-
-  /** @inheritdoc */
-  static get defaultOptions() {
-    return foundry.utils.mergeObject(super.defaultOptions, {
-      template: "systems/dnd5e/templates/advancement/scale-value-flow.hbs"
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  getData() {
-    return foundry.utils.mergeObject(super.getData(), {
-      initial: this.advancement.valueForLevel(this.level - 1)?.display,
-      final: this.advancement.valueForLevel(this.level).display
-    });
-  }
-
 }
