+++ module/documents/_module.mjs
@@ -1,6 +1,7 @@
 // Document Classes
 export {default as ActiveEffect5e} from "./active-effect.mjs";
 export {default as Actor5e} from "./actor/actor.mjs";
+export * as advancement from "./advancement/_module.mjs";
 export {default as Item5e} from "./item.mjs";
 export {default as TokenDocument5e} from "./token.mjs";
