+++ module/documents/advancement/_module.mjs
@@ -1,0 +1,5 @@
+export {default as Advancement} from "./advancement.mjs";
+
+export {default as HitPointsAdvancement} from "./hit-points.mjs";
+export {default as ItemGrantAdvancement} from "./item-grant.mjs";
+export {default as ScaleValueAdvancement} from "./scale-value.mjs";
