+++ module/documents/advancement/hit-points.mjs
@@ -1,13 +1,13 @@
-import Advancement from "../advancement.mjs";
-import AdvancementFlow from "../advancement-flow.mjs";
-import AdvancementConfig from "../advancement-config.mjs";
+import Advancement from "./advancement.mjs";
+import HitPointsConfig from "../../applications/advancement/hit-points-config.mjs";
+import HitPointsFlow from "../../applications/advancement/hit-points-flow.mjs";
 
 /**
  * Advancement that presents the player with the option to roll hit points at each level or select the average value.
  * Keeps track of player hit point rolls or selection for each class level. **Can only be added to classes and each
  * class can only have one.**
  */
-export class HitPointsAdvancement extends Advancement {
+export default class HitPointsAdvancement extends Advancement {
 
   /** @inheritdoc */
   static get metadata() {
@@ -162,109 +162,4 @@ export class HitPointsAdvancement extends Advancement {
     this.updateSource({ [`value.-=${level}`]: null });
     return source;
   }
-}
-
-
-/**
- * Configuration application for hit points.
- */
-export class HitPointsConfig extends AdvancementConfig {
-  /** @inheritdoc */
-  static get defaultOptions() {
-    return foundry.utils.mergeObject(super.defaultOptions, {
-      template: "systems/dnd5e/templates/advancement/hit-points-config.hbs"
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  getData() {
-    return foundry.utils.mergeObject(super.getData(), {
-      hitDie: this.advancement.hitDie
-    });
-  }
-}
-
-
-/**
- * Inline application that presents hit points selection upon level up.
- */
-export class HitPointsFlow extends AdvancementFlow {
-
-  /** @inheritdoc */
-  static get defaultOptions() {
-    return foundry.utils.mergeObject(super.defaultOptions, {
-      template: "systems/dnd5e/templates/advancement/hit-points-flow.hbs"
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  getData() {
-    const source = this.retainedData ?? this.advancement.value;
-    const value = source[this.level];
-
-    // If value is empty, `useAverage` should default to the value selected at the previous level
-    let useAverage = value === "avg";
-    if ( !value ) {
-      const lastValue = source[this.level - 1];
-      if ( lastValue === "avg" ) useAverage = true;
-    }
-
-    return foundry.utils.mergeObject(super.getData(), {
-      isFirstClassLevel: (this.level === 1) && this.advancement.item.isOriginalClass,
-      hitDie: this.advancement.hitDie,
-      dieValue: this.advancement.hitDieValue,
-      data: {
-        value: Number.isInteger(value) ? value : "",
-        useAverage
-      }
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  activateListeners(html) {
-    this.form.querySelector(".averageCheckbox")?.addEventListener("change", event => {
-      this.form.querySelector(".rollResult").disabled = event.target.checked;
-      this.form.querySelector(".rollButton").disabled = event.target.checked;
-      this._updateRollResult();
-    });
-    this.form.querySelector(".rollButton")?.addEventListener("click", async () => {
-      const roll = await this.advancement.actor.rollClassHitPoints(this.advancement.item);
-      this.form.querySelector(".rollResult").value = roll.total;
-    });
-    this._updateRollResult();
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Update the roll result display when the average result is taken.
-   * @protected
-   */
-  _updateRollResult() {
-    if ( !this.form.elements.useAverage?.checked ) return;
-    this.form.elements.value.value = (this.advancement.hitDieValue / 2) + 1;
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  _updateObject(event, formData) {
-    let value;
-    if ( formData.useMax ) value = "max";
-    else if ( formData.useAverage ) value = "avg";
-    else if ( Number.isInteger(formData.value) ) value = parseInt(formData.value);
-
-    if ( value !== undefined ) return this.advancement.apply(this.level, { [this.level]: value });
-
-    this.form.querySelector(".rollResult")?.classList.add("error");
-    const errorType = formData.value ? "Invalid" : "Empty";
-    throw new Advancement.ERROR(game.i18n.localize(`DND5E.AdvancementHitPoints${errorType}Error`));
-  }
-
 }
