+++ module/documents/advancement/item-grant.mjs
@@ -1,14 +1,13 @@
-import Advancement from "../advancement.mjs";
-import AdvancementFlow from "../advancement-flow.mjs";
-import AdvancementConfig from "../advancement-config.mjs";
-
+import Advancement from "./advancement.mjs";
+import ItemGrantConfig from "../../applications/advancement/item-grant-config.mjs";
+import ItemGrantFlow from "../../applications/advancement/item-grant-flow.mjs";
 import ItemGrantConfigurationData from "../../data/advancement/item-grant.mjs";
 
 /**
  * Advancement that automatically grants one or more items to the player. Presents the player with the option of
  * skipping any or all of the items.
  */
-export class ItemGrantAdvancement extends Advancement {
+export default class ItemGrantAdvancement extends Advancement {
 
   /** @inheritdoc */
   static get metadata() {
@@ -123,109 +122,6 @@ export class ItemGrantAdvancement extends Advancement {
     }
     this.updateSource({ "value.-=added": null });
     return { items };
-  }
-
-}
-
-
-/**
- * Configuration application for item grants.
- */
-export class ItemGrantConfig extends AdvancementConfig {
-
-  /** @inheritdoc */
-  static get defaultOptions() {
-    return foundry.utils.mergeObject(super.defaultOptions, {
-      dragDrop: [{ dropSelector: ".drop-target" }],
-      dropKeyPath: "items",
-      template: "systems/dnd5e/templates/advancement/item-grant-config.hbs"
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  getData() {
-    const context = super.getData();
-    context.showSpellConfig = context.configuration.items.map(fromUuidSync).some(i => i.type === "spell");
-    return context;
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  _validateDroppedItem(event, item) {
-    if ( this.advancement.constructor.VALID_TYPES.has(item.type) ) return true;
-    const type = game.i18n.localize(`ITEM.Type${item.type.capitalize()}`);
-    throw new Error(game.i18n.format("DND5E.AdvancementItemTypeInvalidWarning", { type }));
-  }
-}
-
-
-/**
- * Inline application that presents the player with a list of items to be added.
- */
-export class ItemGrantFlow extends AdvancementFlow {
-
-  /** @inheritdoc */
-  static get defaultOptions() {
-    return foundry.utils.mergeObject(super.defaultOptions, {
-      template: "systems/dnd5e/templates/advancement/item-grant-flow.hbs"
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  async getData() {
-    const config = this.advancement.configuration.items;
-    const added = this.retainedData?.items.map(i => foundry.utils.getProperty(i, "flags.dnd5e.sourceId"))
-      ?? this.advancement.value.added;
-    const checked = new Set(Object.values(added ?? {}));
-
-    const items = await Promise.all(config.map(fromUuid));
-    return foundry.utils.mergeObject(super.getData(), {
-      optional: this.advancement.configuration.optional,
-      items: items.reduce((arr, item) => {
-        if ( !item ) return arr;
-        item.checked = added ? checked.has(item.uuid) : true;
-        arr.push(item);
-        return arr;
-      }, [])
-    });
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  activateListeners(html) {
-    super.activateListeners(html);
-    html.find("a[data-uuid]").click(this._onClickFeature.bind(this));
-  }
-
-  /* -------------------------------------------- */
-
-  /**
-   * Handle clicking on a feature during item grant to preview the feature.
-   * @param {MouseEvent} event  The triggering event.
-   * @protected
-   */
-  async _onClickFeature(event) {
-    event.preventDefault();
-    const uuid = event.currentTarget.dataset.uuid;
-    const item = await fromUuid(uuid);
-    item?.sheet.render(true);
-  }
-
-  /* -------------------------------------------- */
-
-  /** @inheritdoc */
-  async _updateObject(event, formData) {
-    const retainedData = this.retainedData?.items.reduce((obj, i) => {
-      obj[foundry.utils.getProperty(i, "flags.dnd5e.sourceId")] = i;
-      return obj;
-    }, {});
-    await this.advancement.apply(this.level, formData, retainedData);
   }
 
 }
